+++ src/common/audio/sound/s_soundinternal.h
@@ -120,7 +120,7 @@ struct FSoundChan : public FISoundChannel
 	float		Volume;
 	int 		EntChannel;	// Actor's sound channel.
 	int			UserData;	// Not used by the engine, the caller can use this to store some additional info.
-	int16_t		Pitch;		// Pitch variation.
+	float		Pitch;		// Pitch variation.
 	int16_t		NearLimit;
 	int8_t		Priority;
 	uint8_t		SourceType;
